chore: move cowSwapTokenToAssetId to cowswap-utils
