+++ src/state/slices/assetsSlice/utils.ts
@@ -1,88 +1,10 @@
-import type {
-  AssetId,
-  AssetReference,
-  ChainId,
-  ChainNamespace,
-  ChainReference,
-} from '@shapeshiftoss/caip'
-import {
-  ASSET_REFERENCE,
-  CHAIN_NAMESPACE,
-  CHAIN_REFERENCE,
-  fromAssetId,
-  fromChainId,
-  isNft,
-  toAssetId,
-} from '@shapeshiftoss/caip'
+import type { AssetId, ChainId } from '@shapeshiftoss/caip'
+import { CHAIN_NAMESPACE, fromAssetId, fromChainId, isNft, toAssetId } from '@shapeshiftoss/caip'
 import type { Transaction } from '@shapeshiftoss/chain-adapters'
 import type { Asset, AssetsByIdPartial } from '@shapeshiftoss/types'
-import { makeAsset } from '@shapeshiftoss/utils'
+import { getNativeFeeAssetReference, makeAsset } from '@shapeshiftoss/utils'
 
 import type { UpsertAssetsPayload } from './assetsSlice'
-
-export const chainIdFeeAssetReferenceMap = (
-  chainNamespace: ChainNamespace,
-  chainReference: ChainReference,
-): AssetReference => {
-  return (() => {
-    switch (chainNamespace) {
-      case CHAIN_NAMESPACE.Utxo:
-        switch (chainReference) {
-          case CHAIN_REFERENCE.BitcoinMainnet:
-            return ASSET_REFERENCE.Bitcoin
-          case CHAIN_REFERENCE.BitcoinCashMainnet:
-            return ASSET_REFERENCE.BitcoinCash
-          case CHAIN_REFERENCE.DogecoinMainnet:
-            return ASSET_REFERENCE.Dogecoin
-          case CHAIN_REFERENCE.LitecoinMainnet:
-            return ASSET_REFERENCE.Litecoin
-          default:
-            throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-        }
-      case CHAIN_NAMESPACE.Evm:
-        switch (chainReference) {
-          case CHAIN_REFERENCE.AvalancheCChain:
-            return ASSET_REFERENCE.AvalancheC
-          case CHAIN_REFERENCE.EthereumMainnet:
-            return ASSET_REFERENCE.Ethereum
-          case CHAIN_REFERENCE.OptimismMainnet:
-            return ASSET_REFERENCE.Optimism
-          case CHAIN_REFERENCE.BnbSmartChainMainnet:
-            return ASSET_REFERENCE.BnbSmartChain
-          case CHAIN_REFERENCE.PolygonMainnet:
-            return ASSET_REFERENCE.Polygon
-          case CHAIN_REFERENCE.GnosisMainnet:
-            return ASSET_REFERENCE.Gnosis
-          case CHAIN_REFERENCE.ArbitrumMainnet:
-            return ASSET_REFERENCE.Arbitrum
-          case CHAIN_REFERENCE.ArbitrumNovaMainnet:
-            return ASSET_REFERENCE.ArbitrumNova
-          case CHAIN_REFERENCE.BaseMainnet:
-            return ASSET_REFERENCE.Base
-          default:
-            throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-        }
-      case CHAIN_NAMESPACE.CosmosSdk:
-        switch (chainReference) {
-          case CHAIN_REFERENCE.CosmosHubMainnet:
-            return ASSET_REFERENCE.Cosmos
-          case CHAIN_REFERENCE.ThorchainMainnet:
-            return ASSET_REFERENCE.Thorchain
-          default:
-            throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-        }
-      case CHAIN_NAMESPACE.Solana:
-        switch (chainReference) {
-          case CHAIN_REFERENCE.SolanaMainnet:
-            return ASSET_REFERENCE.Solana
-          default:
-            throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-        }
-      default:
-        throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-    }
-  })()
-}
 
 type GetFeeAssetByChainId = (
   assetsById: AssetsByIdPartial,
@@ -100,7 +22,7 @@ export const getFeeAssetByChainId: GetFeeAssetByChainId = (assetsById, chainId) 
   const feeAssetId = toAssetId({
     chainId,
     assetNamespace: 'slip44',
-    assetReference: chainIdFeeAssetReferenceMap(chainNamespace, chainReference),
+    assetReference: getNativeFeeAssetReference(chainNamespace, chainReference),
   })
   return assetsById[feeAssetId]
 }
@@ -112,7 +34,7 @@ export const getFeeAssetByAssetId: GetFeeAssetByAssetId = (assetsById, assetId) 
     chainNamespace,
     chainReference,
     assetNamespace: 'slip44',
-    assetReference: chainIdFeeAssetReferenceMap(chainNamespace, chainReference),
+    assetReference: getNativeFeeAssetReference(chainNamespace, chainReference),
   })
   return assetsById[feeAssetId]
 }
